+++ version/prop.go
@@ -40,8 +40,6 @@ func OS() string {
 	return runtime.GOOS
 }
 
-var isSandboxedMacOS lazy.SyncValue[bool]
-
 // IsSandboxedMacOS reports whether this process is a sandboxed macOS
 // process (either the app or the extension). It is true for the Mac App Store
 // and macsys (System Extension) version on macOS, and false for
@@ -210,15 +208,17 @@ var getMeta lazy.SyncValue[Meta]
 
 // GetMeta returns version metadata about the current build.
 func GetMeta() Meta {
-	return Meta{
-		MajorMinorPatch: majorMinorPatch(),
-		Short:           Short(),
-		Long:            Long(),
-		GitCommit:       gitCommit(),
-		GitDirty:        gitDirty(),
-		ExtraGitCommit:  extraGitCommitStamp,
-		IsDev:           isDev(),
-		UnstableBranch:  IsUnstableBuild(),
-		Cap:             int(tailcfg.CurrentCapabilityVersion),
-	}
+	return getMeta.Get(func() Meta {
+		return Meta{
+			MajorMinorPatch: majorMinorPatch(),
+			Short:           Short(),
+			Long:            Long(),
+			GitCommit:       gitCommit(),
+			GitDirty:        gitDirty(),
+			ExtraGitCommit:  extraGitCommitStamp,
+			IsDev:           isDev(),
+			UnstableBranch:  IsUnstableBuild(),
+			Cap:             int(tailcfg.CurrentCapabilityVersion),
+		}
+	})
 }
